moved button enums to a seperate type file
