+++ components/buttons/Button.tsx
@@ -1,30 +1,15 @@
 import Link from 'next/link';
 import { twMerge } from 'tailwind-merge';
-
-export enum ButtonSizeType {
-  small = 'small',
-  default = 'default',
-}
-
-export enum ButtonType {
-  submit = 'submit',
-  reset = 'reset',
-  button = 'button',
-}
-
-export enum ButtonIconPositionType {
-  left = 'left',
-  right = 'right',
-}
+import { ButtonIconPosition, ButtonSize, ButtonType } from '@/types/components/buttons/ButtonPropsType';
 
 type IButtonProps = {
   text: string;
   icon?: React.ReactNode;
-  iconPosition?: ButtonIconPositionType;
+  iconPosition?: ButtonIconPosition;
   target?: string;
   bgClassName?: string;
   textClassName?: string;
-  buttonSize?: ButtonSizeType;
+  buttonSize?: ButtonSize;
   type?: ButtonType;
 } & (
   | ({
@@ -41,11 +26,11 @@ type IButtonProps = {
  * @param {ButtonType} props.type - The type of the button. Defaults to 'button'.
  * @param {string} props.target - The target attribute for the anchor tag. Defaults to '_self'.
  * @param {React.ReactNode} props.icon - The icon to be displayed on the button.
- * @param {ButtonIconPositionType} props.iconPosition - The position of the icon. Defaults to 'right'.
+ * @param {ButtonIconPosition} props.iconPosition - The position of the icon. Defaults to 'right'.
  * @param {string} props.className - The class name to be applied to the button.
  * @param {string} props.bgClassName - The class name to be applied to the button's background.
  * @param {string} props.textClassName - The class name to be applied to the button's text.
- * @param {ButtonSizeType} props.buttonSize - The size of the button. Defaults to 'default'.
+ * @param {ButtonSize} props.buttonSize - The size of the button. Defaults to 'default'.
  * @param {string} props.href - The href attribute for the anchor tag.
  * @description The Button component is a reusable button component that can be used to render a button or an anchor tag
  * @description The component accepts button or anchor tag props based on the presence of the href prop.
@@ -57,7 +42,7 @@ export default function Button({
   type = ButtonType.button,
   target = '_self',
   icon,
-  iconPosition = ButtonIconPositionType.right,
+  iconPosition = ButtonIconPosition.right,
   className,
   bgClassName = twMerge('bg-primary-500 hover:bg-primary-400'),
   textClassName = twMerge('text-white'),
@@ -74,16 +59,16 @@ export default function Button({
       <button
         {...(props as React.ButtonHTMLAttributes<HTMLButtonElement>)}
         type={type}
-        className={buttonSize === ButtonSizeType.small ? smallButtonClasses : classNames}
+        className={buttonSize === ButtonSize.small ? smallButtonClasses : classNames}
         data-testid='Button-main'
       >
-        {icon && iconPosition === ButtonIconPositionType.left && (
+        {icon && iconPosition === ButtonIconPosition.left && (
           <span className='mr-2 inline-block' data-testid='Button-icon-left'>
             {icon}
           </span>
         )}
         <span className='inline-block'>{text}</span>
-        {icon && iconPosition === ButtonIconPositionType.right && (
+        {icon && iconPosition === ButtonIconPosition.right && (
           <span className='ml-2 inline-block' data-testid='Button-icon-right'>
             {icon}
           </span>
@@ -98,12 +83,12 @@ export default function Button({
       {...props}
       target={target}
       rel='noopener noreferrer'
-      className={buttonSize === ButtonSizeType.small ? smallButtonClasses : classNames}
+      className={buttonSize === ButtonSize.small ? smallButtonClasses : classNames}
       data-testid='Button-link'
     >
-      {icon && iconPosition === ButtonIconPositionType.left && <span className='mr-2 inline-block'>{icon}</span>}
+      {icon && iconPosition === ButtonIconPosition.left && <span className='mr-2 inline-block'>{icon}</span>}
       <span className='inline-block'>{text}</span>
-      {icon && iconPosition === ButtonIconPositionType.right && <span className='ml-2 inline-block'>{icon}</span>}
+      {icon && iconPosition === ButtonIconPosition.right && <span className='ml-2 inline-block'>{icon}</span>}
     </Link>
   );
 }
