+++ components/buttons/GithubButton.tsx
@@ -1,5 +1,7 @@
 import IconGithub from '../icons/Github';
-import Button, { ButtonIconPositionType, ButtonSizeType } from './Button';
+import Button from './Button';
+import { ButtonIconPosition, ButtonSize } from '@/types/components/buttons/ButtonPropsType';
+
 import type { IButtonDefaultProps } from './types';
 // TODO: add this again when we have i18n
 // import { useTranslation } from '../../lib/i18n'
@@ -13,14 +15,14 @@ interface IGithubButtonProps extends IButtonDefaultProps {
  * @param {string} props.text - The text to display on the button.
  * @param {string} props.href - The href attribute for the anchor tag.
  * @param {string} props.target - The target attribute for the anchor tag.
- * @param {ButtonIconPositionType} props.iconPosition - The position of the icon in the button.
+ * @param {ButtonIconPosition} props.iconPosition - The position of the icon in the button.
  * @param {string} props.className - The class name to be applied to the button.
  */
 export default function GithubButton({
   text = 'githubButton',
   href = 'https://github.com/asyncapi',
   target = '_blank',
-  iconPosition = ButtonIconPositionType.left,
+  iconPosition = ButtonIconPosition.left,
   className,
   inNav
 }: IGithubButtonProps) {
@@ -39,7 +41,7 @@ export default function GithubButton({
       className={className}
       data-testid='Github-button'
       bgClassName='bg-gray-800 hover:bg-gray-700'
-      buttonSize={inNav ? ButtonSizeType.small : ButtonSizeType.default}
+      buttonSize={inNav ? ButtonSize.small : ButtonSize.default}
     />
   );
 }
