+++ components/buttons/ICSFileButton.tsx
@@ -1,5 +1,6 @@
 import IconCalendar from '../icons/Calendar';
-import Button, { ButtonIconPositionType } from './Button';
+import Button from './Button';
+import { ButtonIconPosition } from '@/types/components/buttons/ButtonPropsType';
 // TODO: add this again when we have i18n
 // import { useTranslation } from '../../lib/i18n';
 import type { IButtonDefaultProps } from './types';
@@ -11,14 +12,14 @@ interface IICSFButtonProps extends IButtonDefaultProps {}
  * @param {string} props.text - The text to display on the button.
  * @param {string} props.href - The href attribute for the anchor tag.
  * @param {string} props.target - The target attribute for the anchor tag.
- * @param {ButtonIconPositionType} props.iconPosition - The position of the icon in the button.
+ * @param {ButtonIconPosition} props.iconPosition - The position of the icon in the button.
  * @param {string} props.className - The class name to be applied to the button.
  */
 export default function ICSFButton({
   text = 'icsFileBtn',
   href,
   target = '_blank',
-  iconPosition = ButtonIconPositionType.left,
+  iconPosition = ButtonIconPosition.left,
   className
 }: IICSFButtonProps) {
   // TODO: add this again when we have i18n
