+++ components/buttons/SlackButton.tsx
@@ -1,5 +1,7 @@
 import IconSlack from '../icons/Slack';
-import Button, { ButtonIconPositionType } from './Button';
+import Button from './Button';
+import { ButtonIconPosition } from '@/types/components/buttons/ButtonPropsType';
+
 import type { IButtonDefaultProps } from './types';
 
 interface ISlackButtonProps extends IButtonDefaultProps {}
@@ -9,14 +11,14 @@ interface ISlackButtonProps extends IButtonDefaultProps {}
  * @param {string} props.text - The text to display on the button.
  * @param {string} props.href - The href attribute for the anchor tag.
  * @param {string} props.target - The target attribute for the anchor tag.
- * @param {ButtonIconPositionType} props.iconPosition - The position of the icon in the button.
+ * @param {ButtonIconPosition} props.iconPosition - The position of the icon in the button.
  * @param {string} props.className - The class name to be applied to the button.
  */
 export default function SlackButton({
   text = 'Join on Slack',
   href = '/slack-invite',
   target = '_blank',
-  iconPosition = ButtonIconPositionType.left,
+  iconPosition = ButtonIconPosition.left,
   className
 }: ISlackButtonProps) {
   return (
