+++ components/buttons/SubscribeButton.tsx
@@ -1,5 +1,7 @@
 import IconSubscribe from '../icons/Subscribe';
-import Button, { ButtonIconPositionType } from './Button';
+import Button from './Button';
+import { ButtonIconPosition } from '@/types/components/buttons/ButtonPropsType';
+
 // import { useTranslation } from '../../lib/i18n';
 import type { IButtonDefaultProps } from './types';
 
@@ -10,14 +12,14 @@ interface IGoogleCalendarButtonProps extends IButtonDefaultProps {}
  * @param {string} props.text - The text to display on the button.
  * @param {string} props.href - The href attribute for the anchor tag.
  * @param {string} props.target - The target attribute for the anchor tag.
- * @param {ButtonIconPositionType} props.iconPosition - The position of the icon in the button.
+ * @param {ButtonIconPosition} props.iconPosition - The position of the icon in the button.
  * @param {string} props.className - The class name to be applied to the button.
  */
 export default function GoogleCalendarButton({
   text = 'subscribeBtn',
   href,
   target = '_blank',
-  iconPosition = ButtonIconPositionType.left,
+  iconPosition = ButtonIconPosition.left,
   className
 }: IGoogleCalendarButtonProps) {
   // const { t } = useTranslation('common');
