+++ components/buttons/YoutubeButton.tsx
@@ -1,5 +1,5 @@
 import IconYoutube from '../icons/YouTube';
-import Button, { ButtonIconPositionType } from './Button';
+import Button, { ButtonIconPosition } from './Button';
 import type { IButtonDefaultProps } from './types';
 
 interface IYoutubeButtonProps extends IButtonDefaultProps {}
@@ -9,14 +9,14 @@ interface IYoutubeButtonProps extends IButtonDefaultProps {}
  * @param {string} props.text - The text to display on the button.
  * @param {string} props.href - The href attribute for the anchor tag.
  * @param {string} props.target - The target attribute for the anchor tag.
- * @param {ButtonIconPositionType} props.iconPosition - The position of the icon in the button.
+ * @param {ButtonIconPosition} props.iconPosition - The position of the icon in the button.
  * @param {string} props.className - The class name to be applied to the button.
  */
 export default function YoutubeButton({
   text = 'Watch on YouTube',
   href,
   target = '_blank',
-  iconPosition = ButtonIconPositionType.left,
+  iconPosition = ButtonIconPosition.left,
   className
 }: IYoutubeButtonProps) {
   return (
