+++ components/buttons/types.ts
@@ -1,11 +1,11 @@
 import type { HTMLAttributeAnchorTarget } from 'react';
 
-import type { ButtonIconPositionType } from './Button';
+import type { ButtonIconPosition } from './Button';
 
 export interface IButtonDefaultProps {
   text?: string;
   href?: string;
   target?: HTMLAttributeAnchorTarget;
-  iconPosition?: ButtonIconPositionType;
+  iconPosition?: ButtonIconPosition;
   className?: string;
 }
