+++ types/components/buttons/ButtonPropsType.ts
@@ -1,0 +1,15 @@
+export enum ButtonSize {
+  small = 'small',
+  default = 'default',
+}
+
+export enum ButtonType {
+  submit = 'submit',
+  reset = 'reset',
+  button = 'button',
+}
+
+export enum ButtonIconPosition {
+  left = 'left',
+  right = 'right',
+}
